ttps/3_C2/gunner_c2/requirements.txt
# --- Backend (API + auth) ---
fastapi>=0.110
uvicorn[standard]>=0.27     # includes websockets/httptools/etc
pydantic>=2.6
PyJWT>=2.8
passlib[bcrypt]>=1.7
cryptography>=42.0
python-multipart>=0.0.9

# --- CLI/Utilities ---
colorama>=0.4.6
requests>=2.31
pyperclip>=1.8.2
tqdm==4.67.1

# --- GUI (PyQt5) ---
PyQt5>=5.15.9
qtawesome>=1.3.0
sip>=6.7
